hoisted the encoder, decoder and polymorphic functionality out of `serialization`
