+++ psl/inc/psl/serialization/serializer.hpp
@@ -10,7 +10,6 @@
 #include "psl/serialization/property.hpp"
 #include "psl/template_utils.hpp"
 #include "psl/ustring.hpp"
-#include "psl/string_utils.hpp"
 
 namespace psl::serialization
 {
@@ -307,573 +306,6 @@ namespace psl::serialization
 		};
 	}	 // namespace details
 
-	class decode_from_format : decoder
-	{
-		using codec_t	  = decode_from_format;
-		using container_t = psl::format::container;
-
-	  public:
-		decode_from_format(psl::format::container& container,
-						   psl::format::handle* root,
-						   std::unordered_map<uint64_t, invocable_wrapper_base*> factory = {}) :
-			m_Container(container),
-			m_CollectionStack {{root}}, m_Factory(factory) {};
-
-		template <auto Name, typename T>
-		decode_from_format& operator<<(property<Name, T>& property)
-		{
-			parse(property);
-			return *this;
-		}
-
-		template <typename... Args>
-		void parse(Args&&... args)
-		{
-			(parse_internal(args), ...);
-		}
-
-		// catch-all for normal types
-		template <auto Name, typename T>
-		void parse(property<Name, T>& property)
-		{
-			parse_internal(property.value, property.name());
-		}
-
-
-		// catch-all for normal, and polymorphic types
-		template <auto Name, typename T>
-		void parse(property<Name, T*>& property)
-		{
-			if constexpr(details::is_collection<T>::value)
-			{
-				property.value = create_polymorphic_collection<T>();
-				parse_collection(*property.value, property.name);
-			}
-			else
-			{
-				property.value = new T();
-				parse_internal(*property.value, property.name());
-			}
-		}
-
-		void resolve_references()
-		{
-			for(auto ref : m_ReferenceNodes)
-			{
-				auto value_opt {ref.second->get().as_reference()};
-				if(!value_opt) continue;
-
-				if(auto it = m_ReferenceMap.find(value_opt.value()); it != m_ReferenceMap.end())
-				{
-					std::memcpy((void*)ref.first, &it->second, sizeof(void*));
-				}
-			}
-			for(auto ptr : m_PointerNodes)
-			{
-				auto value_opt {ptr.second->get().as_reference()};
-				if(!value_opt) continue;
-
-				if(auto it = m_ReferenceMap.find(value_opt.value()); it != m_ReferenceMap.end())
-				{
-					std::memcpy((void*)ptr.first, &it->second, sizeof(void*));
-				}
-			}
-		}
-
-	  private:
-		template <typename T>
-		void parse_collection(T& property, std::optional<psl::string8::view> override_name = {})
-		{
-			if(m_CollectionStack.size() > 0)
-			{
-				parse_collection(property,
-								 m_Container.index_of(m_CollectionStack.top()->get(),
-													  (override_name) ? override_name.value() : accessor::name<T>()));
-			}
-			else
-			{
-				parse_collection(property,
-								 m_Container.index_of((override_name) ? override_name.value() : accessor::name<T>()));
-			}
-		}
-
-		template <typename T>
-		void parse_collection(T& property, psl::format::nodes_t index)
-		{
-			if(index == std::numeric_limits<psl::format::nodes_t>::max()) return;
-
-			auto& collection = m_Container[index];
-			m_CollectionStack.push(&collection);
-
-			m_ReferenceMap[&collection] = (std::uintptr_t)&property;
-
-			if constexpr(details::member_function_serialize<codec_t, T>::value &&
-						 !details::function_serialize<codec_t, T>::value)
-			{
-				accessor::serialize(*this, property);
-			}
-			else if constexpr(!details::member_function_serialize<codec_t, T>::value &&
-							  details::function_serialize<codec_t, T>::value)
-			{
-				accessor::serialize_fn(*this, property);
-			}
-			else
-			{
-				static_assert(utility::templates::always_false_v<T>,
-							  "\n\tPlease define one of the following for the serializer:\n"
-							  "\t\t- a member function of the type template<typename S> void serialize(S& s) {};\n"
-							  "\t\t- a function in the namespace 'serialization' of the signature: template<typename "
-							  "S> void serialize(S& s, T& target) {};\n"
-							  "\t\t- if you did declare a member function, but wanted the serialize to be private, "
-							  "please add 'friend class psl::serialization::serializer' to your class");
-			}
-			m_CollectionStack.pop();
-		}
-
-		template <typename T>
-		T* create_polymorphic_collection(std::optional<psl::string8::view> override_name = {})
-		{
-			if(m_CollectionStack.size() > 0)
-			{
-				auto& collection = m_Container.find(m_CollectionStack.top()->get(),
-													(override_name) ? override_name.value() : accessor::name<T>());
-				m_CollectionStack.push(&collection);
-			}
-			else
-			{
-				auto& collection = m_Container.find((override_name) ? override_name.value() : accessor::name<T>());
-				m_CollectionStack.push(&collection);
-			}
-
-			auto& polymorphic_id = m_Container.find(m_CollectionStack.top()->get(), "POLYMORPHIC_ID");
-			T* target			 = nullptr;
-			if(polymorphic_id.exists())
-			{
-				auto value_opt {polymorphic_id.get().as_value_content()};
-
-				uint64_t id = stoull(psl::string8_t(value_opt.value().second));
-				if(auto it = m_Factory.find(id); it != m_Factory.end())
-				{
-					target = (T*)((*it->second)());
-				}
-				else if(auto it = accessor::polymorphic_data().find(id); it != accessor::polymorphic_data().end())
-				{
-					target = (T*)((*it->second->factory)());
-				}
-				else
-				{
-					target = new T();
-				}
-			}
-			else
-			{
-				target = new T();
-			}
-
-			m_CollectionStack.pop();
-			return target;
-		}
-
-		template <typename T>
-		void parse_internal(T& value, psl::string8::view name)
-		{
-			constexpr bool is_range		 = details::is_range<T>::value;
-			constexpr bool is_collection = details::is_collection<typename std::conditional<
-			  is_range,
-			  typename std::remove_pointer<typename utility::binary::get_contained_type<T>::type>::type,
-			  T>::type>::value;
-
-			if constexpr(is_collection && !is_range)
-			{
-				parse_collection(value, name);
-			}
-			else if constexpr(is_collection && is_range)
-			{
-				using contained_t = typename utility::binary::get_contained_type<T>::type;
-
-				size_t size = 0;
-				if(m_CollectionStack.size() > 0)
-				{
-					auto& collection = m_Container.find(m_CollectionStack.top()->get(), name);
-					if(!collection.exists()) return;
-					m_CollectionStack.push(&collection);
-					auto value_opt {collection.get().as_collection()};
-					size = value_opt.value_or(0);
-				}
-				else
-				{
-					auto& collection = m_Container.find(name);
-					if(!collection.exists()) return;
-					m_CollectionStack.push(&collection);
-					auto value_opt {collection.get().as_collection()};
-					size = value_opt.value_or(0);
-				}
-
-				m_ReferenceMap[m_CollectionStack.top()] = (std::uintptr_t)&value;
-				value.clear();
-				value.reserve(size);
-				size_t begin		= m_Container.index_of(m_CollectionStack.top()->get()) + 1u;
-				size_t end			= begin + size;
-				size_t actual_index = 0;
-				for(auto i = begin; i < end; ++i)
-				{
-					if((size_t)(m_Container[psl::format::nodes_t(i)].get().depth()) - 1u ==
-					   (size_t)(m_CollectionStack.top()->get().depth()))
-					{
-						if constexpr(std::is_pointer<contained_t>::value)
-						{
-							using deref_t = typename std::remove_pointer<contained_t>::type;
-							value.emplace_back(
-							  create_polymorphic_collection<deref_t>(utility::to_string(actual_index)));
-							parse_collection(*value[actual_index], (psl::format::nodes_t)(i));
-						}
-						else
-						{
-							value.emplace_back();
-							parse_collection(value[actual_index], (psl::format::nodes_t)(i));
-						}
-						++actual_index;
-					}
-				}
-				m_CollectionStack.pop();
-			}
-			else if constexpr(details::is_keyed_range<T>::value)
-			{
-				size_t size = 0;
-				if(m_CollectionStack.size() > 0)
-				{
-					auto& collection = m_Container.find(m_CollectionStack.top()->get(), name);
-					m_CollectionStack.push(&collection);
-
-					auto value_opt {collection.get().as_collection()};
-					size = value_opt.value_or(0);
-				}
-				else
-				{
-					auto& collection = m_Container.find(name);
-					m_CollectionStack.push(&collection);
-
-					auto value_opt {collection.get().as_collection()};
-					size = value_opt.value_or(0);
-				}
-
-				size_t begin = m_Container.index_of(m_CollectionStack.top()->get()) + 1u;
-				size_t end	 = begin + size;
-				// size_t actual_index = 0;
-
-				static_assert(details::is_keyed_range<T>::value, "never seen");
-
-				for(auto i = begin; i < end; ++i)
-				{
-					if((size_t)(m_Container[psl::format::nodes_t(i)].get().depth()) - 1u ==
-					   (size_t)(m_CollectionStack.top()->get().depth()))
-					{
-						// auto& pair = value.emplace(, typename T::value_type{});
-						auto& sub_val = value[utility::from_string<typename utility::templates::get_key_type<T>::type>(
-						  m_Container[psl::format::nodes_t(i)].get().name())];
-						parse_collection(sub_val, (psl::format::nodes_t)(i));
-					}
-				}
-				m_CollectionStack.pop();
-			}
-			else if constexpr(is_range)
-			{
-				using contained_t = typename utility::binary::get_contained_type<T>::type;
-
-				auto& node = (m_CollectionStack.size() > 0) ? m_Container.find(m_CollectionStack.top()->get(), name)
-															: m_Container.find(name);
-				if(!node.exists()) return;
-				value.clear();
-
-				auto value_opt {node.get().as_value_range_content()};
-				if(value_opt)
-				{
-					auto data_content {std::move(value_opt.value())};
-					value.reserve(data_content.size());
-
-					for(const auto& it : data_content)
-					{
-						if constexpr(std::is_pointer<contained_t>::value)
-						{
-							using deref_t = typename std::remove_pointer<contained_t>::type;
-							value.emplace_back(new deref_t(
-							  utility::from_string<typename std::remove_pointer<contained_t>::type>(it.second)));
-						}
-						else
-						{
-							value.emplace_back(
-							  utility::from_string<typename utility::binary::get_contained_type<T>::type>(it.second));
-						}
-					}
-				}
-				m_ReferenceMap[&node] = (std::uintptr_t)&value;
-			}
-			else
-			{
-				auto& node = (m_CollectionStack.size() > 0) ? m_Container.find(m_CollectionStack.top()->get(), name)
-															: m_Container.find(name);
-				if(!node.exists()) return;
-
-				if(auto value_opt = node.get().as_value_content(); value_opt)
-				{
-					value = utility::from_string<T>(value_opt.value().second);
-				}
-				m_ReferenceMap[&node] = (std::uintptr_t)&value;
-			}
-		}
-
-		psl::format::container& m_Container;
-		std::stack<psl::format::handle*> m_CollectionStack;
-
-		// reference resolving
-		std::unordered_map<std::uintptr_t, psl::format::handle*> m_ReferenceNodes;
-		std::unordered_map<std::uintptr_t, psl::format::handle*> m_PointerNodes;
-		std::unordered_map<psl::format::handle*, std::uintptr_t> m_ReferenceMap;
-
-		// polymorphic override
-		std::unordered_map<uint64_t, invocable_wrapper_base*> m_Factory;
-	};
-
-	class encode_to_format : encoder
-	{
-		using codec_t	  = encode_to_format;
-		using container_t = psl::format::container;
-
-	  public:
-		encode_to_format(psl::format::container& container, psl::format::handle* root) :
-			m_Container(container), m_CollectionStack {{root}} {};
-
-		template <auto Name, typename T>
-		encode_to_format& operator<<(property<Name, T>& property)
-		{
-			parse(property);
-			return *this;
-		}
-
-		template <typename... Props>
-		void parse(Props&&... props)
-		{
-			(parse(props), ...);
-		}
-
-		template <auto Name, typename T>
-		void parse(property<Name, T>& property)
-		{
-			parse(property.value, property.name());
-		}
-
-		template <auto Name, typename T>
-		void parse(property<Name, T*>& property)
-		{
-			parse(*property.value, property.name());
-		}
-
-		void resolve_references()
-		{
-			for(auto ref : m_ToBeResolvedReferenceMap)
-			{
-				if(auto target = m_ReferenceMap.find(ref.first); target != m_ReferenceMap.end())
-				{
-					m_Container.set_reference(ref.second->get(), target->second->get());
-				}
-			}
-		}
-
-	  private:
-		template <typename T>
-		void parse_collection(T& property, std::optional<psl::string8::view> override_name = {})
-		{
-			auto& collection = m_Container.add_collection(
-			  m_CollectionStack.top()->get(), (override_name) ? override_name.value() : accessor::name<T>());
-			m_CollectionStack.push(&collection);
-			m_ReferenceMap[(std::uintptr_t)&property] = &collection;
-			if constexpr(details::member_function_serialize<codec_t, T>::value)
-			{
-				accessor::serialize(*this, property);
-			}
-			else if constexpr(details::function_serialize<codec_t, T>::value)
-			{
-				accessor::serialize_fn(*this, property);
-			}
-			else
-			{
-				static_assert(utility::templates::always_false_v<T>,
-							  "\n\tPlease define one of the following for the serializer:\n"
-							  "\t\t- a member function of the type template<typename S> void serialize(S& s) {};\n"
-							  "\t\t- a function in the namespace 'serialization' of the signature: template<typename "
-							  "S> void serialize(S& s, T& target) {};\n"
-							  "\t\t- if you did declare a member function, but wanted the serialize to be private, "
-							  "please add 'friend class psl::serialization::serializer' to your class");
-			}
-			m_CollectionStack.pop();
-		}
-		template <typename T>
-		void parse(T& value, psl::string8::view name)
-		{
-			constexpr bool is_range = details::is_range<T>::value;
-			using contained_t		= typename utility::binary::get_contained_type<T>::type;
-
-			constexpr bool is_collection = details::is_collection<
-			  typename std::conditional<is_range, typename std::remove_pointer<contained_t>::type, T>::type>::value;
-
-
-			if constexpr(is_collection && !is_range)
-			{
-				parse_collection(value, name);
-			}
-			else if constexpr(is_collection && is_range)
-			{
-				auto& collection = m_Container.add_collection(m_CollectionStack.top()->get(), name);
-				m_ReferenceMap[(std::uintptr_t)&value] = &collection;
-				m_CollectionStack.push(&collection);
-				size_t index = 0;
-
-				if constexpr(std::is_pointer<contained_t>::value)
-				{
-					for(contained_t& it : value)
-					{
-						parse_collection(*it, utility::to_string(index++));
-					}
-				}
-				else
-				{
-					for(contained_t& it : value)
-					{
-						parse_collection(it, utility::to_string(index++));
-					}
-				}
-				m_CollectionStack.pop();
-			}
-			else if constexpr(details::is_keyed_range<T>::value)
-			{
-				auto& collection = m_Container.add_collection(m_CollectionStack.top()->get(), name);
-				m_ReferenceMap[(std::uintptr_t)&value] = &collection;
-				m_CollectionStack.push(&collection);
-				for(auto& pair : value)
-				{
-					static_assert(details::is_collection<typename utility::templates::get_value_type<T>::type>::value,
-								  "can only use associative containers when their value is a collection");
-					parse_collection(
-					  pair.second, utility::to_string<typename utility::templates::get_key_type<T>::type>(pair.first));
-				}
-				m_CollectionStack.pop();
-			}
-			else if constexpr(is_range)
-			{
-				psl::string8_t intermediate;
-				std::vector<size_t> locations;
-				std::vector<psl::string8::view> res;
-				res.reserve(value.size());
-				if constexpr(std::is_pointer<contained_t>::value)
-				{
-					for(const auto& it : value)
-					{
-						intermediate.append(utility::to_string<typename std::remove_pointer<contained_t>::type>(*it));
-						locations.push_back(intermediate.size());
-					}
-				}
-				else
-				{
-					for(const auto& it : value)
-					{
-						intermediate.append(utility::to_string<contained_t>(it));
-						locations.push_back(intermediate.size());
-					}
-				}
-
-				size_t previous = 0u;
-				for(const auto& loc : locations)
-				{
-					res.emplace_back(&intermediate[previous], loc - previous);
-					previous = loc;
-				}
-				m_Container.add_value_range(m_CollectionStack.top()->get(), name, res);
-				m_ReferenceMap[(std::uintptr_t)&value] = &m_Container[(psl::format::nodes_t)(m_Container.size() - 1u)];
-			}
-			else
-			{
-				m_Container.add_value(m_CollectionStack.top()->get(), name, utility::to_string<contained_t>(value));
-				m_ReferenceMap[(std::uintptr_t)&value] = &m_Container[(psl::format::nodes_t)(m_Container.size() - 1u)];
-			}
-		}
-
-
-		psl::format::container& m_Container;
-		std::stack<psl::format::handle*> m_CollectionStack;
-
-		std::unordered_map<std::uintptr_t, psl::format::handle*> m_ReferenceMap;
-		std::unordered_map<std::uintptr_t, psl::format::handle*> m_ToBeResolvedReferenceMap;
-	};
-
-	static property<"POLYMORPHIC_ID", psl::string8_t> p;
-
-	/// \brief contains the polymorphic lambda constructor for your polymorphic type
-	///
-	/// This will be used by the deserializer to create the correct type for the format node.
-	/// All polymorphic types are internally mapped to an ID (generated from a name you give).
-	/// These values are saved in the format, and on deserialization are used to figure out how
-	/// to construct the object in the format.
-	/// Aside from the fact you need to register your polymorphic types (once), there should be no further
-	/// interaction with this class.
-	template <typename T>
-	class polymorphic final : public polymorphic_base
-	{
-	  public:
-		polymorphic()
-		{
-			auto& pData = accessor::polymorphic_data();
-			if(pData.find(ID) != pData.end())
-			{
-				LOG_FATAL("Encountered duplicate Polymorphic ID in the serialization: ", ID);
-				exit(-1);
-			}
-
-
-			static polymorphic_data_t data;
-
-			data.id = ID;
-			auto lambda {[]() { return new T(); }};
-			data.factory = new invocable_wrapper<decltype(lambda)>(std::forward<decltype(lambda)>(lambda));
-
-			data.encode = [](void* this_, encode_to_format& s) {
-				T* t	= static_cast<T*>(this_);
-				p.value = utility::to_string(accessor::polymorphic_id(t));
-				s.parse(p);
-				accessor::serialize_directly<encode_to_format, T>(s, *t);
-				// t->serialize(s);
-			};
-			data.decode = [](void* this_, decode_from_format& s) {
-				accessor::serialize_directly<decode_from_format, T>(s, *static_cast<T*>(this_));
-				// static_assert(false, "use of undefined codec");
-			};
-			pData[ID] = &data;
-		}
-
-		virtual ~polymorphic() {};
-		uint64_t PolymorphicID() const override { return accessor::id<T>(); };
-		static constexpr uint64_t ID {accessor::id<T>()};
-	};
-
-	template <typename Base, typename... Rest>
-	static const void notify_base(psl::string8_t name, uint64_t ID)
-	{
-		accessor::polymorphic_data()[accessor::id<Base>()]->derived.emplace_back(name, ID);
-		if constexpr(sizeof...(Rest) > 0)
-		{
-			notify_base<Rest...>(name, ID);
-		}
-	}
-
-	template <typename T, typename... Base>
-	static const uint64_t register_polymorphic()
-	{
-		static psl::serialization::polymorphic<T> polymorphic_container;
-		if constexpr(sizeof...(Base) > 0)
-		{
-			notify_base<Base...>(accessor::name<T>(), accessor::id<T>());
-		}
-		return accessor::id<T>();
-	}
 	class serializer
 	{
 		template <typename T, typename Encoder>
